+++ BootShop/Controllers/Admin/ProductController.cs
@@ -4,12 +4,18 @@ using Microsoft.EntityFrameworkCore;
 
 namespace BootShop.Controllers.Admin
 {
-    public class ProductController : Controller
+    public class ProductController : AdminBaseController
     {
         private BootShopContext context = new BootShopContext();
         [HttpGet]
         public IActionResult Index()
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             ViewBag.Products = this.context.Products.Include(p=>p.Brand).Include(p=>p.Category).Include(p=>p.Subcategory);
             ViewBag.Brands = this.context.Brands;
             ViewBag.Categories = this.context.Categories;
@@ -21,6 +27,12 @@ namespace BootShop.Controllers.Admin
         [HttpPost]
         public IActionResult Index(Product product)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.Products.Add(product);
             this.context.SaveChanges();
 
@@ -35,6 +47,12 @@ namespace BootShop.Controllers.Admin
         [HttpGet]
         public IActionResult Delete(Product product)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.Products.Remove(product);
             this.context.SaveChanges();
 
@@ -44,6 +62,12 @@ namespace BootShop.Controllers.Admin
         [HttpGet]
         public IActionResult Edit(Product product)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             ViewBag.Product = this.context.Products.Find(product.Id);
             ViewBag.Brands = this.context.Brands;
             ViewBag.Categories = this.context.Categories;
@@ -55,6 +79,12 @@ namespace BootShop.Controllers.Admin
         [ActionName("Edit")]
         public IActionResult EditPost(Product product)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.Products.Update(product);
             this.context.SaveChanges();
