+++ BootShop/Controllers/Admin/ProductVariantController.cs
@@ -4,12 +4,18 @@ using Microsoft.EntityFrameworkCore;
 
 namespace BootShop.Controllers.Admin
 {
-    public class ProductVariantController : Controller
+    public class ProductVariantController : AdminBaseController
     {
         private BootShopContext context = new BootShopContext();
         [HttpGet]
         public IActionResult Index()
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             ViewBag.ProductVariants = this.context.ProductVariants.Include(v => v.Product).Include(v => v.Color);
             ViewBag.Products = this.context.Products;
             ViewBag.Colors = this.context.Colors;
@@ -21,6 +27,12 @@ namespace BootShop.Controllers.Admin
         [HttpPost]
         public IActionResult Index(ProductVariant productVariant)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.ProductVariants.Add(productVariant);
             this.context.SaveChanges();
 
@@ -35,6 +47,12 @@ namespace BootShop.Controllers.Admin
         [HttpGet]
         public IActionResult Delete(ProductVariant productVariant)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.ProductVariants.Remove(productVariant);
             this.context.SaveChanges();
 
@@ -44,6 +62,12 @@ namespace BootShop.Controllers.Admin
         [HttpGet]
         public IActionResult Edit(ProductVariant productVariant)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             ViewBag.ProductVariant = this.context.ProductVariants.Find(productVariant.Id);
             ViewBag.Products = this.context.Products;
             ViewBag.Colors = this.context.Colors;
@@ -55,6 +79,12 @@ namespace BootShop.Controllers.Admin
         [ActionName("Edit")]
         public IActionResult EditPost(ProductVariant productVariant)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.ProductVariants.Update(productVariant);
             this.context.SaveChanges();
