+++ BootShop/Controllers/Admin/SizeController.cs
@@ -3,13 +3,19 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace BootShop.Controllers.Admin
 {
-    public class SizeController : Controller
+    public class SizeController : AdminBaseController
     {
         private BootShopContext context = new BootShopContext();
 
         [HttpGet]
         public IActionResult Index()
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             ViewBag.Sizes = this.context.Sizes.ToList();
 
             return View("/Views/Admin/Size.cshtml");
@@ -18,6 +24,12 @@ namespace BootShop.Controllers.Admin
         [HttpPost]
         public IActionResult Index(Size size)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.Sizes.Add(size);
             this.context.SaveChanges();
             ViewBag.Sizes = this.context.Sizes.ToList();
@@ -28,6 +40,12 @@ namespace BootShop.Controllers.Admin
         [HttpGet]
         public IActionResult Delete(Size size)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.Sizes.Remove(size);
             this.context.SaveChanges();
