+++ BootShop/Controllers/Admin/SubcategoryController.cs
@@ -4,12 +4,18 @@ using Microsoft.EntityFrameworkCore;
 
 namespace BootShop.Controllers.Admin
 {
-    public class SubcategoryController : Controller
+    public class SubcategoryController : AdminBaseController
     {
         private BootShopContext context = new();
         [HttpGet]
         public IActionResult Index()
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             ViewBag.Subcategories = this.context.Subcategories.Include(s => s.Category); //najoinuje data z category https://learn.microsoft.com/en-us/ef/ef6/querying/related-data 
             ViewBag.Categories = this.context.Categories;
 
@@ -19,6 +25,12 @@ namespace BootShop.Controllers.Admin
         [HttpPost]
         public IActionResult Index(Subcategory subcategory)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.Subcategories.Add(subcategory);
             this.context.SaveChanges();
 
@@ -31,6 +43,12 @@ namespace BootShop.Controllers.Admin
         [HttpGet]
         public IActionResult Delete(Subcategory subcategory)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.Subcategories.Remove(subcategory);
             this.context.SaveChanges();
 
@@ -40,6 +58,12 @@ namespace BootShop.Controllers.Admin
         [HttpGet]
         public IActionResult Edit(Subcategory subcategory)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             ViewBag.Subcategory = this.context.Subcategories.Where(s => s.Id == subcategory.Id).Include(s => s.Category).Single();
 
             return View("/Views/Admin/SubcategoryEdit.cshtml");
@@ -49,11 +73,17 @@ namespace BootShop.Controllers.Admin
         [ActionName("Edit")]
         public IActionResult EditPost(Subcategory subcategory)
         {
+            RedirectToActionResult? checkloginResult = this.checkLogin();
+            if (checkloginResult != null)
+            {
+                return checkloginResult;
+            }
+
             this.context.Subcategories.Update(subcategory);
             this.context.SaveChanges();
 
             ViewBag.Subcategory = this.context.Subcategories.Where(s => s.Id == subcategory.Id).Include(s => s.Category).Single();
-            
+
             return View("/Views/Admin/SubcategoryEdit.cshtml");
         }
     }
